test: ApiHelper 공통화
