+++ src/test/java/com/camping/admin/steps/AdminSteps.java
@@ -1,6 +1,6 @@
 package com.camping.admin.steps;
 
-import com.camping.common.support.AdminApiHelper;
+import com.camping.common.support.ApiHelper;
 import com.camping.common.support.CommonContext;
 import io.cucumber.core.options.CurlOption.HttpMethod;
 import io.cucumber.java.en.Then;
@@ -17,14 +17,14 @@ public class AdminSteps {
 
     @When("admin 컨테이너에 요청을 보낸다")
     public void admin컨테이너에요청을보낸다() {
-        CommonContext.lastResponse = AdminApiHelper.request(HttpMethod.GET, "/", null)
+        CommonContext.lastResponse = ApiHelper.adminRequest(HttpMethod.GET, "/", null)
                 .then().log().all()
                 .extract().response();
     }
 
     @When("상품 목록 조회를 요청한다")
     public void 상품목록조회를요청한다() {
-        CommonContext.lastResponse = AdminApiHelper.request(HttpMethod.GET, "admin/products", null)
+        CommonContext.lastResponse = ApiHelper.adminRequest(HttpMethod.GET, "admin/products", null)
                 .then().log().all()
                 .extract().response();
     }
